Items in Inventory
Picking up items will make the item sprite appear in the inventory overlayed on top of the item box. If an item is consumable (only quest or misc items), then the item gets removed, along with the overlayed sprite, and all items will 'move left' to fill in the gap and the display will be updated.